# back-end/docs/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "CalmWave API",
    "description": "API para plataforma de processamento de áudio com IA. Documentação completa de todas as rotas disponíveis.",
    "version": "1.0.0",
    "contact": {
      "name": "CalmWave Team",
      "email": "support@calmwave.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5000/api",
      "description": "Development Server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "User": {
        "type": "object",
        "properties": {
          "id": { "type": "integer" },
          "name": { "type": "string" },
          "email": { "type": "string" },
          "profile_photo_url": { "type": "string" },
          "created_at": { "type": "string", "format": "date-time" },
          "last_access": { "type": "string", "format": "date-time" },
          "active": { "type": "boolean" },
          "account_type": { "type": "string", "enum": ["free", "premium", "admin"] },
          "settings": {
            "type": "object",
            "properties": {
              "dark_mode": { "type": "boolean" },
              "notifications_enabled": { "type": "boolean" },
              "auto_process_audio": { "type": "boolean" },
              "audio_quality": { "type": "string", "enum": ["low", "normal", "high"] }
            }
          }
        }
      },
...
    }
  },
  "paths": {
...
    "/users": {
      "get": {
        "tags": ["Users"],
        "summary": "Listar usuários",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    "total": { "type": "integer" },
                    "pages": { "type": "integer" },
                    "page": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/users/{user_id}": {
      "get": {
        "tags": ["Users"],
        "summary": "Obter usuário específico",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Dados do usuário",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "put": {
        "tags": ["Users"],
        "summary": "Atualizar usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "active": { "type": "boolean" },
                  "account_type": { "type": "string" },
                  "profile_photo_url": { "type": "string" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Users"],
        "summary": "Deletar usuário",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Usuário deletado com sucesso" }
        }
      }
    },
...
    "/admin/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Listar usuários (Admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": { "type": "integer", "default": 1 }
          },
          {
            "name": "per_page",
            "in": "query",
            "schema": { "type": "integer", "default": 20 }
          },
          {
            "name": "search",
            "in": "query",
            "schema": { "type": "string" }
          },
          {
            "name": "account_type",
            "in": "query",
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de usuários",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": { "type": "array", "items": { "$ref": "#/components/schemas/User" } },
                    "total": { "type": "integer" },
                    "pages": { "type": "integer" },
                    "page": { "type": "integer" }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["Admin"],
        "summary": "Criar novo usuário (Admin)",
        "security": [{ "bearerAuth": [] }],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["name", "email", "password"],
                "properties": {
                  "name": { "type": "string" },
                  "email": { "type": "string", "format": "email" },
                  "password": { "type": "string", "format": "password" },
                  "account_type": { "type": "string", "enum": ["free", "premium", "admin"] }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Usuário criado com sucesso",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      }
    },
    "/admin/users/{user_id}": {
      "put": {
        "tags": ["Admin"],
        "summary": "Atualizar usuário (Admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": { "type": "string" },
                  "account_type": { "type": "string" },
                  "active": { "type": "boolean" }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Usuário atualizado",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/User" }
              }
            }
          }
        }
      },
      "delete": {
        "tags": ["Admin"],
        "summary": "Deletar usuário (Admin)",
        "security": [{ "bearerAuth": [] }],
        "parameters": [
          {
            "name": "user_id",
            "in": "path",
            "required": true,
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": { "description": "Usuário deletado com sucesso" }
        }
      }
    },
    "/admin/reports/overview": {
      "get": {
        "tags": ["Admin"],
        "summary": "Relatório geral do sistema",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dados do relatório",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "users": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "active": { "type": "integer" },
                        "admins": { "type": "integer" },
                        "premium": { "type": "integer" },
                        "free": { "type": "integer" }
                      }
                    },
                    "audios": {
                      "type": "object",
                      "properties": {
                        "total": { "type": "integer" },
                        "processed": { "type": "integer" },
                        "processed_pct": { "type": "number" },
                        "transcribed": { "type": "integer" },
                        "favorite": { "type": "integer" }
                      }
                    },
                    "metrics": {
                      "type": "object"
                    },
                    "today": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/reports/users": {
      "get": {
        "tags": ["Admin"],
        "summary": "Relatório de usuários",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dados do relatório",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "daily_registrations": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "day": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "top_users": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": { "type": "integer" },
                          "name": { "type": "string" },
                          "email": { "type": "string" },
                          "audio_count": { "type": "integer" }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/admin/reports/audios": {
      "get": {
        "tags": ["Admin"],
        "summary": "Relatório de áudios",
        "security": [{ "bearerAuth": [] }],
        "responses": {
          "200": {
            "description": "Dados do relatório",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "daily_uploads": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "day": { "type": "string" },
                          "count": { "type": "integer" }
                        }
                      }
                    },
                    "processing": {
                      "type": "object"
                    },
                    "storage": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}